requirements.txt
appdirs==1.4.4
cachetools==3.1.1
certifi==2020.12.5
chardet==4.0.0
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==1.7.1
distlib==0.3.1
enum34==1.1.10
ffmpeg==1.4
filelock==3.0.12
futures==3.1.1
gdown==3.12.2
google-auth==1.28.1
googleapis-common-protos==1.52.0
idna==2.10
importlib-metadata==2.1.1
importlib-resources==3.3.1
ipaddress==1.0.17
keyring==10.1
keyrings.alt==1.3
packaging==20.9
pathlib2==2.3.5
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pydub==0.24.1
pygobject==3.27.0
pyparsing==2.4.7
PySocks==1.7.1
pytz==2021.1
pyxdg==0.25
requests==2.25.1
rsa==4.5
scandir==1.10.0
SecretStorage==2.3.1
singledispatch==3.6.1
six==1.15.0
SpeechRecognition==3.8.1
tqdm==4.59.0
typing==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.2
zipp==1.2.0
